1. Changed AmountOfMoney from decimal to double 2. Updated migration 3. Added Serilog package as a Logger 4. Added additional validation for AmountOfMoney value 5. Added a new unit test

+++ PaymentSystem/PaymentSystem.Contract/Models/BankAccount.cs
@@ -9,7 +9,7 @@ namespace PaymentSystem.Contract.Models
 
         public DateTime DateCreated { get; set; }
 
-        public decimal AmountOfMoney { get; set; }
+        public double AmountOfMoney { get; set; }
 
         public IEnumerable<Card> PaymentCards { get; set; }
     }
