+++ PaymentSystem/PaymentSystem.Infrastructure/Transaction/ITransactionService.cs
@@ -4,6 +4,6 @@ namespace PaymentSystem.Infrastructure.Transaction
 {
     public interface ITransactionService
     {
-        Task TransferMoneyAsync(string senderCardNumber, string RecipientCardId, decimal amountOfMoney);
+        Task TransferMoneyAsync(string senderCardNumber, string RecipientCardId, double amountOfMoney);
     }
 }
